requirements.txt
discord.py[voice]>=2.7.0
python-dotenv>=1.0.0
yt-dlp>=2024.0.0
PyNaCl>=1.5.0
flask>=3.0.0
flask-cors>=4.0.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.27.0
psutil>=5.9.0
gtts>=2.5.0